refactor to views folder
