+++ src/views/Deposit.js
@@ -10,13 +10,13 @@ import {
 } from '@material-ui/core';
 import HowToVoteIcon from '@material-ui/icons/HowToVote';
 import WithdrawIcon from '@material-ui/icons/RemoveCircle';
-import { Page } from '../../components';
-import Header from '../../components/Header/Header';
-import LoadingWeb3 from '../../components/LoadingWeb3/LoadingWeb3';
-import EllipsisLoader from '../../components/EllipsisLoader/EllipsisLoader';
-import useRouter from '../../utils/useRouter';
-import useWeb3Connect from '../../utils/useWeb3Connect';
-import { useRedirectHomeIfNoEthAccount } from '../../utils/useCommonUtils';
+import { Page } from '../components';
+import Header from '../components/Header/Header';
+import LoadingWeb3 from '../components/LoadingWeb3/LoadingWeb3';
+import EllipsisLoader from '../components/EllipsisLoader/EllipsisLoader';
+import useRouter from '../utils/useRouter';
+import useWeb3Connect from '../utils/useWeb3Connect';
+import { useRedirectHomeIfNoEthAccount } from '../utils/useCommonUtils';
 import { useForm } from 'react-hook-form';
 
 const BN = require('bn.js');
