--- src/views/SubmitProposal/components/index.js
@@ -1,7 +1,0 @@
-// export { default as Intro } from './Intro';
-// export { default as Newsletter } from './Newsletter';
-// export { default as Details } from './Details';
-// export { default as Team } from './Team';
-// export { default as Partners } from './Partners';
-// export { default as Footer } from './Footer';
-// export { default as PadeleeRegistration } from './PadeleeRegistration';
